Tentative fix for i3 focus_follows_mouse issue.
In i3 with focus_follows_mouse enabled, wmfocus would seemingly fail to put the
focus on windows that did not contain the mouse pointer. By adding in delays in
the process and checking the logs, it seems like there is "mouse enter" event
that occurs in the desktop windows *after* the hint window is
unmapped/destroyed. Since the hint windows were destroyed after focus was placed
on the desktop window, this resulted in focus being placed in the window under
the mouse cursor even if this was not the window selected in wmfocus.

The fix here unmaps the hint windows *before* focus is placed on the selected
desktop window. This seems to work OK but I'm not sure it's the best solution.
There could be a race like the following: Once the hint windows are unmapped,
i3/X replaces the mouse cursor and generates the "mouse enter" event either
before or after wm::focus is called. Maybe this is not a huge issue.